+++ gcdapi/runtime.go
@@ -148,7 +148,7 @@ type RuntimeConsoleAPICalledEvent struct {
 		Args               []*RuntimeRemoteObject `json:"args"`                 // Call arguments.
 		ExecutionContextId int                    `json:"executionContextId"`   // Identifier of the context where the call was made.
 		Timestamp          float64                `json:"timestamp"`            // Call timestamp.
-		StackTrace         *RuntimeStackTrace     `json:"stackTrace,omitempty"` // Stack trace captured when the call was made.
+		StackTrace         *RuntimeStackTrace     `json:"stackTrace,omitempty"` // Stack trace captured when the call was made. The async stack chain is automatically reported for the following call types: `assert`, `error`, `trace`, `warning`. For other types the async call chain can be retrieved using `Debugger.getStackTrace` and `stackTrace.parentId` field.
 		Context            string                 `json:"context,omitempty"`    // Console context descriptor for calls on non-default console context (not console.*): 'anonymous#unique-logger-id' for call on unnamed context, 'name#unique-logger-id' for call on named context.
 	} `json:"Params,omitempty"`
 }