+++ gcdapi/tracing.go
@@ -36,6 +36,7 @@ type TracingTracingCompleteEvent struct {
 	Method string `json:"method"`
 	Params struct {
 		Stream            string `json:"stream,omitempty"`            // A handle of the stream that holds resulting trace data.
+		TraceFormat       string `json:"traceFormat,omitempty"`       // Trace data format of returned stream. enum values: json, proto
 		StreamCompression string `json:"streamCompression,omitempty"` // Compression format of returned stream. enum values: none, gzip
 	} `json:"Params,omitempty"`
 }
@@ -146,6 +147,8 @@ type TracingStartParams struct {
 	BufferUsageReportingInterval float64 `json:"bufferUsageReportingInterval,omitempty"`
 	// Whether to report trace events as series of dataCollected events or to save trace to a stream (defaults to `ReportEvents`).
 	TransferMode string `json:"transferMode,omitempty"`
+	// Trace data format to use. This only applies when using `ReturnAsStream` transfer mode (defaults to `json`). enum values: json, proto
+	StreamFormat string `json:"streamFormat,omitempty"`
 	// Compression format to use. This only applies when using `ReturnAsStream` transfer mode (defaults to `none`) enum values: none, gzip
 	StreamCompression string `json:"streamCompression,omitempty"`
 	//
@@ -162,14 +165,16 @@ func (c *Tracing) StartWithParams(v *TracingStartParams) (*gcdmessage.ChromeResp
 // options - Tracing options
 // bufferUsageReportingInterval - If set, the agent will issue bufferUsage events at this interval, specified in milliseconds
 // transferMode - Whether to report trace events as series of dataCollected events or to save trace to a stream (defaults to `ReportEvents`).
+// streamFormat - Trace data format to use. This only applies when using `ReturnAsStream` transfer mode (defaults to `json`). enum values: json, proto
 // streamCompression - Compression format to use. This only applies when using `ReturnAsStream` transfer mode (defaults to `none`) enum values: none, gzip
 // traceConfig -
-func (c *Tracing) Start(categories string, options string, bufferUsageReportingInterval float64, transferMode string, streamCompression string, traceConfig *TracingTraceConfig) (*gcdmessage.ChromeResponse, error) {
+func (c *Tracing) Start(categories string, options string, bufferUsageReportingInterval float64, transferMode string, streamFormat string, streamCompression string, traceConfig *TracingTraceConfig) (*gcdmessage.ChromeResponse, error) {
 	var v TracingStartParams
 	v.Categories = categories
 	v.Options = options
 	v.BufferUsageReportingInterval = bufferUsageReportingInterval
 	v.TransferMode = transferMode
+	v.StreamFormat = streamFormat
 	v.StreamCompression = streamCompression
 	v.TraceConfig = traceConfig
 	return c.StartWithParams(&v)